Made: pre-commit lint-staed, Avatars module, data caching, Balance module, logs

+++ src/modules/auth/auth.service.ts
@@ -1,4 +1,9 @@
-import { Injectable, UnauthorizedException } from '@nestjs/common'
+import {
+  Injectable,
+  Logger,
+  NotFoundException,
+  UnauthorizedException,
+} from '@nestjs/common'
 import * as bcrypt from 'bcryptjs'
 
 import { UsersService } from '../users/users.service'
@@ -16,6 +21,8 @@ import { ERROR_MESSAGES } from 'src/common/consts/error-messages.const'
 
 @Injectable()
 export class AuthService {
+  private readonly logger = new Logger(AuthService.name)
+
   constructor(
     private readonly jwtTokensSevice: JwtTokensSevice,
     private readonly userService: UsersService,
@@ -24,6 +31,8 @@ export class AuthService {
   ) {}
 
   async signUp(userDto: CreateUserDto): Promise<TokensResDto> {
+    this.logger.log('🔍 Beginning of signing up')
+
     const createdUser = await this.userService.createUser(userDto)
     const createdUserJson = createdUser.toJSON()
 
@@ -31,8 +40,8 @@ export class AuthService {
       id: createdUserJson.id,
       login: createdUserJson.login,
       age: createdUserJson.age,
-      description: createdUserJson.description,
       email: createdUserJson.email,
+      description: createdUserJson.description,
       createdAt: createdUserJson.createdAt,
       updatedAt: createdUserJson.updatedAt,
     }
@@ -53,10 +62,14 @@ export class AuthService {
       tokens.refreshToken,
     )
 
+    this.logger.log('✅ Signing up was successful')
+
     return tokens
   }
 
   async signIn(signInDto: SignInDto): Promise<TokensResDto> {
+    this.logger.log('🔍 Beginning of signing in')
+
     const user = await this.userRepository.getUserByEmail(signInDto.email)
     const userJson = user.toJSON()
 
@@ -64,6 +77,7 @@ export class AuthService {
       !user ||
       !(await bcrypt.compare(signInDto.password, user.toJSON().password))
     ) {
+      this.logger.error('❌ Invalid credentials')
       throw new UnauthorizedException(ERROR_MESSAGES.INVALID_CREDITIONALS)
     }
 
@@ -95,32 +109,41 @@ export class AuthService {
       tokens.refreshToken,
     )
 
+    this.logger.log('✅ Signing in was successful')
+
     return tokens
   }
 
   async logout(userJwtPayload: IJwtPayload): Promise<GetUserResDto> {
+    this.logger.log('🔍 Beginning of logging out')
     const currentSession = await this.sessionRepository.getSessionById(
       userJwtPayload.id,
     )
 
     if (!currentSession) {
+      this.logger.error('❌ Invalid token')
       throw new UnauthorizedException(ERROR_MESSAGES.INVALID_TOKEN)
+    }
+
+    const user = await this.userRepository.getUserByIdWithoutPassword(
+      userJwtPayload.id,
+    )
+
+    if (!user) {
+      this.logger.error('❌ User not found')
+      throw new NotFoundException(ERROR_MESSAGES.NO_USER_WITH_THIS_ID)
     }
 
     await this.sessionRepository.destroySession(userJwtPayload.id)
 
-    return {
-      id: userJwtPayload.id,
-      login: userJwtPayload.login,
-      age: userJwtPayload.age,
-      description: userJwtPayload.description,
-      email: userJwtPayload.email,
-      createdAt: userJwtPayload.createdAt,
-      updatedAt: userJwtPayload.updatedAt,
-    }
+    this.logger.log('✅ Logging out was successful')
+
+    return user.toJSON() as GetUserResDto
   }
 
   async updateToken(refreshToken: string) {
+    this.logger.log('🔍 Beginning of updating token')
+
     const refreshTokenPayload = this.jwtTokensSevice.verifyToken(
       refreshToken,
       process.env.JWT_REFRESH_SECRET || '',
@@ -133,6 +156,7 @@ export class AuthService {
       !currentSession ||
       currentSession.toJSON().refreshToken !== refreshToken
     ) {
+      this.logger.error('❌ Invalid token')
       throw new UnauthorizedException(ERROR_MESSAGES.INVALID_TOKEN)
     }
 
@@ -140,11 +164,13 @@ export class AuthService {
       id: refreshTokenPayload.id,
       login: refreshTokenPayload.login,
       age: refreshTokenPayload.age,
-      description: refreshTokenPayload.description,
       email: refreshTokenPayload.email,
+      description: refreshTokenPayload.description,
       createdAt: refreshTokenPayload.createdAt,
       updatedAt: refreshTokenPayload.updatedAt,
     }
+
+    this.logger.log('✅ Updating token was successful')
 
     return this.jwtTokensSevice.generateToken('access', accessTokenPayload)
   }
